requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
attrs==25.3.0
    # via
    #   outcome
    #   trio
aws-lambda-powertools==3.19.0
    # via newton-llm (pyproject.toml)
beautifulsoup4==4.13.5
    # via newton-llm (pyproject.toml)
certifi==2025.8.3
    # via selenium
dnspython==2.7.0
    # via pymongo
exceptiongroup==1.3.0
    # via
    #   trio
    #   trio-websocket
h11==0.16.0
    # via wsproto
idna==3.10
    # via trio
jmespath==1.0.1
    # via aws-lambda-powertools
outcome==1.3.0.post0
    # via
    #   trio
    #   trio-websocket
pydantic==2.11.7
    # via
    #   newton-llm (pyproject.toml)
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via newton-llm (pyproject.toml)
pymongo==4.14.1
    # via newton-llm (pyproject.toml)
pysocks==1.7.1
    # via urllib3
python-dotenv==1.1.1
    # via pydantic-settings
selenium==4.35.0
    # via newton-llm (pyproject.toml)
sniffio==1.3.1
    # via trio
sortedcontainers==2.4.0
    # via trio
soupsieve==2.8
    # via beautifulsoup4
structlog==25.4.0
    # via newton-llm (pyproject.toml)
trio==0.30.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.12.2
    # via selenium
typing-extensions==4.14.1
    # via
    #   aws-lambda-powertools
    #   beautifulsoup4
    #   exceptiongroup
    #   pydantic
    #   pydantic-core
    #   selenium
    #   structlog
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3[socks]==2.5.0
    # via selenium
websocket-client==1.8.0
    # via selenium
wsproto==1.2.0
    # via trio-websocket
